Make review action entrypoint fail closed
