+++ .github/actions/review-coverage/ci-review-coverage.mjs
@@ -9,60 +9,9 @@
 // Run from the JavaScript action in this directory, or locally:
 //   node .github/actions/review-coverage/ci-review-coverage.mjs --event <payload.json> [--mode enforce]
 
-import { appendFileSync, readFileSync, realpathSync } from 'node:fs';
-import { fileURLToPath } from 'node:url';
-import { COVERAGE_REQUIRED, isTrivialChange, reportedCrossModelReviews } from './reviewGate.mjs';
-
-const NON_MEMBER_ASSOCIATIONS = new Set([
-	'COLLABORATOR',
-	'CONTRIBUTOR',
-	'FIRST_TIME_CONTRIBUTOR',
-	'FIRST_TIMER',
-	'MANNEQUIN',
-	'NONE',
-]);
-
-/** The whole decision, pure. `pr` is the pull_request object from the Actions event
- *  payload. Returns { pass, exempt, summary, detail } — `pass` already accounts for mode. */
-export function evaluateCiCoverage(pr, { mode = 'report', required = COVERAGE_REQUIRED } = {}) {
-	const login = String(pr?.user?.login ?? '');
-	const assoc = String(pr?.author_association ?? '');
-	const body = String(pr?.body ?? '');
-	const { count, families } = reportedCrossModelReviews(body);
-	const plural = count === 1 ? 'review' : 'reviews';
-	const reported = `${count} cross-model ${plural} reported${families.length ? ` (${families.join(', ')})` : ''}`;
-	const coverage = count >= required ? reported : `${reported} — policy asks for ${required}`;
-
-	// The Human-Review-Need footer is evidence a prepush review RAN; staleness means
-	// commits landed after the last reviewed head (HEG step 14's most-skipped step).
-	const footer = [...body.matchAll(/Human-Review-Need:\s*(\d+)(?:[^@\n]*@\s*([0-9a-f]{6,40}))?/gi)].at(-1);
-	const head = String(pr?.head?.sha ?? '').toLowerCase();
-	const footerNote = !footer
-		? 'no Human-Review-Need footer'
-		: !footer[2]
-			? `Human-Review-Need: ${footer[1]} @ unpinned sha`
-			: head.startsWith(footer[2].toLowerCase())
-				? `Human-Review-Need: ${footer[1]} @ head`
-				: `Human-Review-Need footer is STALE (reviewed @ ${footer[2].slice(0, 7)}, head is ${head.slice(0, 7)})`;
-
-	// The triviality exemption requires the size fields to actually be present: a payload
-	// missing additions/deletions would otherwise read as 0+0 and silently exempt everything.
-	const sized = Number.isFinite(pr?.additions) && Number.isFinite(pr?.deletions);
-	const exempt =
-		login.endsWith('[bot]') || pr?.user?.type === 'Bot'
-			? 'bot author'
-			: NON_MEMBER_ASSOCIATIONS.has(assoc)
-				? `author is not an org member (${assoc}) — always human review`
-				: sized && isTrivialChange(pr?.additions, pr?.deletions)
-					? 'trivial change (≤2 lines)'
-					: pr?.draft
-						? 'draft — checked again at ready-for-review'
-						: '';
-	const compliant = count >= required;
-	const pass = mode !== 'enforce' || Boolean(exempt) || compliant;
-	const summary = exempt ? `exempt: ${exempt}` : coverage;
-	return { pass, exempt, compliant, count, families, summary, detail: `${coverage}; ${footerNote}` };
-}
+import { appendFileSync, readFileSync } from 'node:fs';
+import { evaluateCiCoverage } from './evaluateCiCoverage.mjs';
+import { COVERAGE_REQUIRED } from './reviewGate.mjs';
 
 function arg(name, fallback = '') {
 	const i = process.argv.indexOf(`--${name}`);
@@ -98,8 +47,6 @@ function main(mode) {
 		}
 	}
 	console.log(`review-coverage [${mode}]: ${r.exempt ? r.summary : r.detail}`);
-	// report mode stays green, but an under-reported PR still gets the count on the PR
-	// surface as a warning annotation rather than only in the job summary
 	if (r.pass && !r.exempt && !r.compliant)
 		console.error(
 			`::warning::${r.detail} — the dispatch gate will block at review time if the fleet's review finds issues`
@@ -112,34 +59,21 @@ function main(mode) {
 	}
 }
 
-function invokedDirectly() {
+const mode = arg('mode', process.env.INPUT_MODE || process.env.REVIEW_COVERAGE_MODE || 'report').toLowerCase();
+if (mode !== 'report' && mode !== 'enforce') {
+	console.error(`::error::review-coverage: unknown mode '${mode}' (report|enforce)`);
+	process.exitCode = 1;
+} else {
 	try {
-		return realpathSync(process.argv[1]) === realpathSync(fileURLToPath(import.meta.url));
-	} catch {
-		return false;
-	}
-}
-if (invokedDirectly()) {
-	const mode = arg('mode', process.env.INPUT_MODE || process.env.REVIEW_COVERAGE_MODE || 'report').toLowerCase();
-	if (mode !== 'report' && mode !== 'enforce') {
-		// a typo'd mode must be loud — 'enfroce' silently meaning report is the bad direction
-		console.error(`::error::review-coverage: unknown mode '${mode}' (report|enforce)`);
-		process.exitCode = 1;
-	} else {
-		try {
-			main(mode);
-		} catch (error) {
-			// plumbing failures follow the mode: informational stays green, but an enforcing
-			// check that cannot read its payload must not silently wave PRs through
-			console.error(
-				`review-coverage: ${error.message}${mode === 'enforce' ? '' : ' (passing — report mode fails only on policy)'}`
-			);
-			if (mode === 'enforce') {
-				console.error(
-					`::error::review-coverage could not evaluate this PR (${error.message}) — enforce mode fails closed`
-				);
-				process.exitCode = 1;
-			}
+		main(mode);
+	} catch (error) {
+		const message = error instanceof Error ? error.message : String(error);
+		console.error(
+			`review-coverage: ${message}${mode === 'enforce' ? '' : ' (passing — report mode fails only on policy)'}`
+		);
+		if (mode === 'enforce') {
+			console.error(`::error::review-coverage could not evaluate this PR (${message}) — enforce mode fails closed`);
+			process.exitCode = 1;
 		}
 	}
 }
